API cleanup
As I write more code with this, it becomes more clear that the
right structure for a Compound is a map[string]Payload, because
otherwise the tag and the key in the compound could differ,
which is obviously wrong. Shuffle things around a bit and
add Get[tagtype] functions which operate on payloads, rather than
being methods on them.